# api.yml
openapi: 3.0.3
info:
  title: RSU/ESOP Platform API
  description: |
    API for managing equity compensation (RSUs/stock options) for startups.
    Features include pool management, grant administration, vesting calculations, and employee portal access.
    
    Key platform features:
    - Single ordinary-share class
    - Fixed DECIMAL(12,3) precision for all share & monetary amounts
    - 12-month cliff, 48-month total vesting
    - Roles: Admin, Employee
    - Pool & PPS management with vesting engine
    - Daily vesting batch job
    - Soft delete on all tables
  version: '1.0.0'
  contact:
    name: Platform Support
    email: support@rsu-platform.com
servers:
  - url: 'https://api.rsu-platform.com/api'
    description: Production API server
  - url: 'https://staging.rsu-platform.com/api'
    description: Staging server for testing
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0-issued JWT Access Token.
        Obtained via Auth0 Universal Login flow.
        Contains standard claims plus potential custom claims for role/permissions.
        Example claims expected by backend middleware:
        - sub: Auth0 User ID (used as auth0_user_id)
        - aud: API Identifier (audience)
        - iss: Issuer URL
        - exp: Expiration timestamp
        - iat: Issued at timestamp
        - scope: Permissions (if using RBAC)
        - Custom claims for role/tenant_id (if configured in Auth0 Rules/Actions)
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: 'INVALID_GRANT'
            message:
              type: string
              description: Human-readable error message
              example: 'The specified grant does not exist or is inaccessible'
            details:
              type: object
              description: Field-specific validation errors
              example:
                grant_id: 'Invalid UUID format'
                share_amount: 'Must be greater than 0.001'
                status: 'Must be one of: active, inactive'
    Success:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data, schema varies by endpoint
...
    User:
      type: object
      required:
        - user_id
        - tenant_id
        - auth0_user_id
        - email
        - name
        - role
        - status
        - created_at
      properties:
        user_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        auth0_user_id:
          type: string
          description: The unique user identifier from Auth0 (subject claim).
          example: 'auth0|6638aaabbccddeeff112233'
        email:
          type: string
          format: email
          example: 'user@example.com'
        name:
          type: string
          example: 'John Doe'
        role:
          type: string
          enum: [admin, employee]
          example: 'admin'
        status:
          type: string
          enum: [active, inactive]
          example: 'active'
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        deleted_at:
          type: string
          format: date-time
          nullable: true
          example: null
          description: 'Soft delete timestamp, if applicable'
    UserCreate:
      type: object
      required:
        - email
        - name
        - role
      properties:
        email:
          type: string
          format: email
          description: User's email address. Must match the one they will use with Auth0.
          example: 'user@example.com'
        name:
          type: string
          example: 'John Doe'
        role:
          type: string
          enum: [admin, employee]
          example: 'admin'
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          example: 'John D. Doe'
        status:
          type: string
          enum: [active, inactive]
          example: 'inactive'
        role:
          type: string
          enum: [admin, employee]
          example: 'employee'
...
    PaginatedResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            type: object
        pagination:
          type: object
          required:
            - total_items
            - total_pages
            - current_page
            - limit
          properties:
            total_items:
              type: integer
              example: 45
            total_pages:
              type: integer
              example: 3
            current_page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            next_page:
              type: integer
              nullable: true
              example: 2
            prev_page:
              type: integer
              nullable: true
              example: null
...
  parameters:
...
    UserIdParam:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the user
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page for pagination
    StatusParam:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, inactive]
      description: Filter by status
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      summary: Create user application record
      description: |
        Creates a user record within the application, associating an email with a role.
        The actual user authentication is handled by Auth0.
        This endpoint is typically used by an Admin to pre-register a user or set their role
        before the user logs in for the first time via Auth0.
        Requires Admin role.
      security:
        - bearerAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    
    get:
      summary: List users for the tenant
      description: Retrieve paginated list of users for the current tenant. Requires Admin role.
      security:
        - bearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedResponse'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/{user_id}:
    get:
      summary: Get user details by application ID
      description: Retrieve specific user details by their internal application user_id. Requires Admin role.
      security:
        - bearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update user details by application ID
      description: Update user details (e.g., name, status, role). Requires Admin role.
      security:
        - bearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete user by application ID (Soft Delete)
      description: Soft delete a user record within the application. Requires Admin role.
      security:
        - bearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...